Add AbstractFile-backed FlatClasspath implementation.
It's not being used yet because we have to hook into PathResolver
and construct those objects for each classpath element.

Also, this is just to get us started with the whole idea of replacing
classpath implementation. I expect we'll get specialized versions
for JarFlatClasspath and DirectoryFlatClasspath that are optimized
for the fact that we are dealing with a jar file or a directory.