Introduce FlatClasspath.list method.
It allows to scan given package in one pass which comes very handy
when we list ClasspathEntries in a directory.

The implementation in PackageLoaderUsingFlatClasspath is a big
involved because of irregularity related to empty package.
Long-term, those special cases should be removed or handled somewhere
else.

Also, in DirectoryFlatClasspath switched to constructing PlainFile
directly in order to avoid expensive `isFile` call.