Implement incremental compiler classpath compatibility hack.
Incremental compiler interface depends on ClassPath.findClass method.
Since there's no common interface between the old ClassPath and
the new FlatClasspath we are stuck.

The only way I found to be compatible with incremental compiler is to
make the old classpath implementation aware of the new one, and
dynamically dispatch to FlatClasspath if it's enabled. The old classpath
is all lazy so if the only call to the old classpath is coming
from incremental compiler it'll be dispatched to the new classpath
and nothing in old classpath will get forced.